fix: wait for file close when saving tar file
We build a tar file with image contents, and save it to a temp folder in
the file system. This fix waits for file close event before resolving
the function's promise, in order to make sure that the caller of the
function has all the expected contents inside the tar file before
extracting it.